+++ src/core/camera.h
@@ -6,16 +6,16 @@
 
 struct Camera {
   Camera() {};
-  Camera(Vec3 e, Vec3 g, Vec3 u, double vfov, RenderParams& params);
+  Camera(Vec3 e, Vec3 g, Vec3 u, float vfov, RenderParams& params);
 
   virtual Ray getRay(int i, int j, RNG& rng);
   virtual Ray getRay(int i, int j);
   
   Vec3 e;
-  double du, dv, wl, wt;
+  float du, dv, wl, wt;
   Matrix4 C2W, W2C;
 
   bool DOF = false;
-  double aperture, focus_dist;
+  float aperture, focus_dist;
 };
