+++ src/core/ray.h
@@ -4,12 +4,12 @@
 
 struct Ray {
   Ray(Vec3 p, Vec3 d): p(p), d(d), tMax(INFINITY) {}
-  Ray(Vec3 p, Vec3 d, double tMax): p(p), d(d), tMax(tMax) {}
+  Ray(Vec3 p, Vec3 d, float tMax): p(p), d(d), tMax(tMax) {}
 
-  inline Vec3 at(double lambda) const { return p + lambda * d; }
+  inline Vec3 at(float lambda) const { return p + lambda * d; }
 
   Vec3 p;
   Vec3 d;
-  double tMax = INFINITY;
+  float tMax = INFINITY;
 };
