+++ src/core/tile_manager.h
@@ -15,10 +15,10 @@ struct TileManager {
     int sy = params.get<int>("height");
 
     // Compute window bounds in pixels
-    window.xBegin = floor(sx * params.get<double>("xBegin"));
-    window.yBegin = floor(sy * params.get<double>("yBegin"));
-    window.xEnd = ceil(sx * params.get<double>("xEnd"));
-    window.yEnd = ceil(sy * params.get<double>("yEnd"));
+    window.xBegin = floor(sx * params.get<float>("xBegin"));
+    window.yBegin = floor(sy * params.get<float>("yBegin"));
+    window.xEnd = ceil(sx * params.get<float>("xEnd"));
+    window.yEnd = ceil(sy * params.get<float>("yEnd"));
 
     tileSize = params.get<int>("tileSize");
 
@@ -26,15 +26,15 @@ struct TileManager {
     int numPixelsInWindow = (window.xEnd - window.xBegin) * 
                             (window.yEnd - window.yBegin);
     int numPixelsInTile = tileSize * tileSize;
-    int maxTiles = ceil( numPixelsInWindow / (double) numPixelsInTile);
+    int maxTiles = ceil( numPixelsInWindow / (float) numPixelsInTile);
 
     // If this is lower than some threshold, then change the tile size
     if (maxTiles < MIN_TILES) {
-      tileSize = ceil( sqrt(numPixelsInWindow / (double) MIN_TILES) );
+      tileSize = ceil( sqrt(numPixelsInWindow / (float) MIN_TILES) );
     }
 
-    xTiles = ceil((window.xEnd - window.xBegin) / (double)tileSize);
-    yTiles = ceil((window.yEnd - window.yBegin) / (double)tileSize);
+    xTiles = ceil((window.xEnd - window.xBegin) / (float)tileSize);
+    yTiles = ceil((window.yEnd - window.yBegin) / (float)tileSize);
     numTiles = xTiles * yTiles;
   }
