+++ src/materials/orennayar.cc
@@ -1,9 +1,9 @@
 #include <materials/orennayar.h>
 
-OrenNayar::OrenNayar(double roughness, Colour col) 
+OrenNayar::OrenNayar(float roughness, Colour col) 
   : BSDF(col) {
   // roughness = Radians(sigma);
-  double sigma2 = roughness * roughness;
+  float sigma2 = roughness * roughness;
   A = 1.f - (sigma2 / (2.f * (sigma2 + 0.33f)));
   B = 0.45f * sigma2 / (sigma2 + 0.09f);
 }
@@ -18,19 +18,19 @@ Colour OrenNayar::eval(HitRec &rec) {
   // std::cout << "----------------  END -----------------" << std::endl;
 
 
-  double sinThetaI = sinTheta(wi);
-  double sinThetaO = sinTheta(wo);
+  float sinThetaI = sinTheta(wi);
+  float sinThetaO = sinTheta(wo);
   // Compute cosine term of Oren-Nayar model
-  double maxCos = 0;
+  float maxCos = 0;
   if (sinThetaI > 1e-4 && sinThetaO > 1e-4) {
-      double sinPhiI = sinPhi(wi), cosPhiI = cosPhi(wi);
-      double sinPhiO = sinPhi(wo), cosPhiO = cosPhi(wo);
-      double dCos = cosPhiI * cosPhiO + sinPhiI * sinPhiO;
+      float sinPhiI = sinPhi(wi), cosPhiI = cosPhi(wi);
+      float sinPhiO = sinPhi(wo), cosPhiO = cosPhi(wo);
+      float dCos = cosPhiI * cosPhiO + sinPhiI * sinPhiO;
       maxCos = max(0, dCos);
   }
 
   // Compute sine and tangent terms of Oren-Nayar model
-  double sinAlpha, tanBeta;
+  float sinAlpha, tanBeta;
   if (absCosTheta(wi) > absCosTheta(wo)) {
       sinAlpha = sinThetaO;
       tanBeta = sinThetaI / absCosTheta(wi);
@@ -44,10 +44,10 @@ Colour OrenNayar::eval(HitRec &rec) {
 
 Colour OrenNayar::sample(HitRec &rec, RNG& rng) {
   rec.wi = rng.randomVectorCosineHemisphere(rec.n);
-  double cosTheta = dot(rec.n, rec.wi);
+  float cosTheta = dot(rec.n, rec.wi);
   return (cosTheta * eval(rec)) / pdf(rec);
 }
 
-double OrenNayar::pdf(HitRec &rec) {
+float OrenNayar::pdf(HitRec &rec) {
   return dot(rec.wi, rec.n) * INV_PI;
 }
