+++ src/materials/transmissive.h
@@ -7,18 +7,18 @@
 #include <core/bsdf.h>
 
 struct Transmissive : BSDF {
-  Transmissive(double ior, Colour col) : BSDF(col), ior(ior) { 
+  Transmissive(float ior, Colour col) : BSDF(col), ior(ior) { 
     specular = true; 
   }
-  Transmissive(double ior, Texture *tx) : BSDF(tx), ior(ior) { 
+  Transmissive(float ior, Texture *tx) : BSDF(tx), ior(ior) { 
     specular = true; 
   }
   Colour eval(HitRec& rec);
   Colour sample(HitRec& rec, RNG& rng);
   Colour reflect(HitRec& rec);
-  double pdf(HitRec& rec);
+  float pdf(HitRec& rec);
 
   // Index of refraction
-  double ior;
+  float ior;
 };
