+++ src/objects/cube.h
@@ -5,6 +5,6 @@
 struct Cube : Object {
   Cube(BSDF *mat) : Object(mat) { bounds = AABB(1); };
   bool hit(Ray& r, HitRec &rec);
-  Vec3 sample(double *pdf, RNG& rng);
+  Vec3 sample(float *pdf, RNG& rng);
 };
