+++ src/util/image.h
@@ -10,16 +10,16 @@ struct Image {
   Image(char const *fname);
   ~Image();
   
-  void save(char const *fname, bool gammaCorrect=true, double exposure=1.0);
+  void save(char const *fname, bool gammaCorrect=true, float exposure=1.0);
   
   void set(int i, int j, Colour col);
   void splat(int i, int j, Colour col);
   
   Colour get(int i, int j);
-  Colour get(double u, double v);
+  Colour get(float u, float v);
   
   int sx, sy;
-  double *data;
+  float *data;
 };
 
 struct RTImageList {
