+++ test/test_perlin_orig.cc
@@ -7,9 +7,9 @@ int main() {
   Perlin perlin = Perlin(pscale);
   for (int i = 0; i < 512; i++) {
     for (int j = 0; j < 512; j++) {
-      double u = i/(512.0/pscale), v = j/(512.0/pscale), w = 1; 
+      float u = i/(512.0/pscale), v = j/(512.0/pscale), w = 1; 
       Vec3 p = Vec3(u, v, w);
-      double scale = 10;
+      float scale = 10;
       Vec3 col = 0.5*(1 + sin(scale * p.z + 10*perlin.turb(p)));
       im.set(i, j, col);
     }
